+++ Services/TicketStore.Web/Controllers/EventsController.cs
@@ -31,6 +31,10 @@ namespace TicketStore.Web.Controllers
                 return new BadRequestObjectResult("Request should contains merchantId parameter");
             }
 
+            var resultTemp = _db.Events
+                .Where(e => e.MerchantId == merchantId)
+                .ToList();
+            
             var result = _db.Events
                 .Where(e => e.MerchantId == merchantId)
                 .OrderBy(e => e.Time)
